DHFPROD-6681 - Refactoring the existing e2e scripts
