+++ marklogic-data-hub-central/ui/e2e/cypress/integration/curation/curate/addMergeStepToFlow.spec.tsx
@@ -33,11 +33,9 @@ describe("Add Merge step to a flow", () => {
     cy.deleteFlows("mergeE2ETest", "mergeE2ETestRun");
     cy.resetTestUser();
   });
-  it("Navigating to curate tab", () => {
+  it("Navigating to Customer Merge tab", () => {
     cy.waitUntil(() => toolbar.getCurateToolbarIcon()).click();
     cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-  });
-  it("Open Customer entity", () => {
     curatePage.toggleEntityTypeId("Customer");
     curatePage.selectMergeTab("Customer");
   });
@@ -52,7 +50,7 @@ describe("Add Merge step to a flow", () => {
     cy.waitForAsyncRequest();
     curatePage.verifyStepNameIsVisible(mergeStep);
   });
-  it("Create merge step with duplicate name", () => {
+  it("Create merge step with duplicate name and verify duplicate name modal is displayed", () => {
     cy.waitUntil(() => curatePage.addNewStep()).click();
     createEditStepDialog.stepNameInput().type(mergeStep);
     createEditStepDialog.stepDescriptionInput().type("merge order step example");
@@ -60,13 +58,11 @@ describe("Add Merge step to a flow", () => {
     createEditStepDialog.setQueryInput("test");
     createEditStepDialog.saveButton("merging").click();
     cy.waitForAsyncRequest();
-  });
-  it("Verify duplicate name modal is displayed", () => {
     loadPage.duplicateStepErrorMessage();
     loadPage.confirmationOptions("OK").click();
     loadPage.duplicateStepErrorMessageClosed();
   });
-  it("Add the Merge step to new flow and validate the step was added", () => {
+  it("Add the Merge step to new flow and Run the step(new)", () => {
     curatePage.addToNewFlow("Customer", mergeStep);
     cy.findByText("New Flow").should("be.visible");
     runPage.setFlowName(flowName1);
@@ -74,33 +70,27 @@ describe("Add Merge step to a flow", () => {
     loadPage.confirmationOptions("Save").click();
     cy.waitForAsyncRequest();
     cy.verifyStepAddedToFlow("Merge", mergeStep);
-  });
-  it("Run the merge step and verify success message(new)", () => {
     runPage.runStep(mergeStep).click();
     cy.waitForAsyncRequest();
     cy.verifyStepRunResult("success", "Merging", mergeStep);
     tiles.closeRunMessage();
   });
-  it("Delete the step", () => {
+  it("Delete the step and Navigate back to merge tab", () => {
     runPage.deleteStep(mergeStep).click();
     loadPage.confirmationOptions("Yes").click();
     cy.waitForAsyncRequest();
-  });
-  it("Navigate back to merge tab", () => {
     cy.waitUntil(() => toolbar.getCurateToolbarIcon()).click();
     cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
     curatePage.toggleEntityTypeId("Customer");
     curatePage.selectMergeTab("Customer");
   });
-  it("Add the Merge step to an existing flow and validate the step was added", () => {
+  it("Add the Merge step to an existing flow and Run the step(existing)", () => {
     curatePage.openExistingFlowDropdown("Customer", mergeStep);
     curatePage.getExistingFlowFromDropdown(flowName1).click();
     curatePage.addStepToFlowConfirmationMessage();
     curatePage.confirmAddStepToFlow(mergeStep, flowName1);
     cy.waitForAsyncRequest();
     cy.verifyStepAddedToFlow("Merge", mergeStep);
-  });
-  it("Run the merge step and verify success message(existing)", () => {
     runPage.runStep(mergeStep).click();
     cy.waitForAsyncRequest();
     cy.verifyStepRunResult("success", "Merging", mergeStep);
@@ -130,12 +120,10 @@ describe("Add Merge step to a flow", () => {
     cy.verifyStepRunResult("success", "Merging", mergeStep);
     tiles.closeRunMessage();
   });
-  it("Delete the merge step", () => {
+  it("Delete the merge step and Navigating to merge tab", () => {
     runPage.deleteStep(mergeStep).click();
     loadPage.confirmationOptions("Yes").click();
     cy.waitForAsyncRequest();
-  });
-  it("Navigating to merge tab", () => {
     cy.waitUntil(() => toolbar.getCurateToolbarIcon()).click();
     cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
     curatePage.toggleEntityTypeId("Customer");
