+++ marklogic-data-hub-central/ui/e2e/cypress/integration/curation/curate/mappingWithInterceptorsAndCustomHook.spec.tsx
@@ -9,38 +9,34 @@ import loadPage from "../../../support/pages/load";
 import browsePage from "../../../support/pages/browse";
 import curatePage from "../../../support/pages/curate";
 import runPage from "../../../support/pages/run";
-import detailPage from "../../../support/pages/detail";
 import LoginPage from "../../../support/pages/login";
 import "cypress-wait-until";
 
-describe("Mapping", () => {
+const flowName = "orderFlow";
+const loadStep = "loadOrder";
+const mapStep = "mapOrder";
 
-  beforeEach(() => {
+describe("Create and verify load steps, map step and flows with interceptors & custom hook", () => {
+  before(() => {
     cy.visit("/");
     cy.contains(Application.title);
     cy.loginAsTestUserWithRoles("hub-central-flow-writer", "hub-central-mapping-writer", "hub-central-load-writer").withRequest();
     LoginPage.postLogin();
-    cy.waitUntil(() => toolbar.getCurateToolbarIcon()).click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
   });
-
+  beforeEach(() => {
+    cy.loginAsTestUserWithRoles("hub-central-flow-writer", "hub-central-mapping-writer", "hub-central-load-writer").withRequest();
+  });
   afterEach(() => {
     cy.resetTestUser();
   });
-
   after(() => {
     cy.loginAsDeveloper().withRequest();
-    cy.deleteSteps("ingestion", "loadOrder", "loadOrderCustomHeader");
-    cy.deleteSteps("mapping", "mapOrder", "mapOrderCustomHeader");
-    cy.deleteSteps("mapping", "mapCustomer");
-    cy.deleteFlows("orderFlow", "orderCustomHeaderFlow");
+    cy.deleteSteps("ingestion", "loadOrder");
+    cy.deleteSteps("mapping", "mapOrder", "mapCustomer");
+    cy.deleteFlows("orderFlow");
+    cy.resetTestUser();
   });
-
-  it("can create load step with interceptors & custom hook, can create mapping step with interceptors & custom hook, can create new flow, run both steps, and verify interceptors & custom hooks", () => {
-    const flowName = "orderFlow";
-    const loadStep = "loadOrder";
-    const mapStep = "mapOrder";
-    // create load step
+  it("Create load step", () => {
     toolbar.getLoadToolbarIcon().click();
     cy.waitUntil(() => loadPage.stepName("ingestion-step").should("be.visible"));
     loadPage.addNewButton("card").click();
@@ -54,7 +50,8 @@ describe("Mapping", () => {
 
     loadPage.confirmationOptions("Save").click({force: true});
     cy.findByText(loadStep).should("be.visible");
-
+  });
+  it("Edit load step", () => {
     // Open step settings and switch to Advanced tab
     loadPage.editStepInCardView(loadStep).click({force: true});
     loadPage.switchEditAdvanced().click();
@@ -68,8 +65,8 @@ describe("Mapping", () => {
 
     advancedSettingsDialog.saveSettings(loadStep).click();
     advancedSettingsDialog.saveSettings(loadStep).should("not.be.visible");
-
-    //verify load step with duplicate name cannot be created
+  });
+  it("Verify load step with duplicate name cannot be created", () => {
     loadPage.addNewButton("card").click();
     loadPage.saveButton().should("be.enabled");
     loadPage.stepNameInput().type(loadStep);
@@ -77,8 +74,8 @@ describe("Mapping", () => {
     loadPage.duplicateStepErrorMessage();
     loadPage.confirmationOptions("OK").click();
     loadPage.duplicateStepErrorMessageClosed();
-
-    // add step to new flow
+  });
+  it("Add step to new flow and Run", () => {
     loadPage.addStepToNewFlow(loadStep);
     cy.findByText("New Flow").should("be.visible");
     runPage.editSave().should("be.enabled");
@@ -86,14 +83,13 @@ describe("Mapping", () => {
     runPage.setFlowDescription(`${flowName} description`);
     loadPage.confirmationOptions("Save").click();
     cy.verifyStepAddedToFlow("Load", loadStep);
-
     //Run the ingest with JSON
     runPage.runStep(loadStep).click();
     cy.uploadFile("input/10259.json");
     cy.verifyStepRunResult("success", "Ingestion", loadStep);
     tiles.closeRunMessage();
-
-    // create mapping step
+  });
+  it("Create mapping step", () => {
     toolbar.getCurateToolbarIcon().click();
     cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
     curatePage.toggleEntityTypeId("Order");
@@ -102,20 +98,17 @@ describe("Mapping", () => {
     createEditMappingDialog.setMappingDescription("An order mapping with custom interceptors");
     createEditMappingDialog.setSourceRadio("Query");
     createEditMappingDialog.setQueryInput(`cts.collectionQuery(['${loadStep}'])`);
-
     //verify advanced setting modifications during creation
     loadPage.switchEditAdvanced().click();
     // add interceptor to map step
     advancedSettingsDialog.setStepInterceptor("curateTile/orderDateInterceptor");
-
     createEditMappingDialog.saveButton().click({force: true});
     cy.waitForAsyncRequest();
     cy.waitUntil(() => curatePage.dataPresent().should("be.visible"));
-
     //verify that step details automatically opens after step creation
     curatePage.verifyStepDetailsOpen(mapStep);
-
-    //Validate xpath expressions are blank by default
+  });
+  it("Validate xpath expressions are blank by default", () => {
     curatePage.xpathExpression("orderId").should("have.value", "");
     curatePage.xpathExpression("address").should("have.value", "");
     curatePage.xpathExpression("city").should("have.value", "");
@@ -127,8 +120,8 @@ describe("Mapping", () => {
     curatePage.xpathExpression("discount").should("have.value", "");
     curatePage.xpathExpression("shipRegion").should("have.value", "");
     curatePage.xpathExpression("shippedDate").should("have.value", "");
-
-    //Map source to entity
+  });
+  it("Map source to entity and Test the mappings", () => {
     sourceToEntityMap.setXpathExpressionInput("orderId", "OrderID");
     sourceToEntityMap.setXpathExpressionInput("address", "/");
     sourceToEntityMap.setXpathExpressionInput("city", "ShipCity");
@@ -142,8 +135,7 @@ describe("Mapping", () => {
     sourceToEntityMap.setXpathExpressionInput("shippedDate", "ShippedDate");
     curatePage.dataPresent().should("be.visible");
     sourceToEntityMap.expandCollapseEntity().click();
-
-    //Test the mappings
+    // Test the mappings
     cy.waitUntil(() => sourceToEntityMap.testMap().should("be.enabled"));
     sourceToEntityMap.testMap().click();
     sourceToEntityMap.validateMapValues("orderId", "10259");
@@ -157,12 +149,11 @@ describe("Mapping", () => {
     sourceToEntityMap.validateMapValues("discount", "0");
     sourceToEntityMap.validateMapValues("shipRegion", "region1\nregion4\n");
     sourceToEntityMap.validateMapValues("shippedDate", "1996-07-17T00:28:30");
-
     //close modal
     cy.get("body").type("{esc}");
-
     curatePage.verifyStepNameIsVisible(mapStep);
-
+  });
+  it("Edit Map step", () => {
     // Open step details and switch to Advanced tab in step settings
     curatePage.openStepDetails(mapStep);
     cy.waitUntil(() => curatePage.dataPresent().should("be.visible"));
@@ -170,18 +161,15 @@ describe("Mapping", () => {
     sourceToEntityMap.validateMapValues("orderId", "10259");
     sourceToEntityMap.stepSettingsLink().click();
     curatePage.switchEditAdvanced().click();
-
     //interceptor should already be set during creation
     cy.findByLabelText("interceptors-expand").trigger("mouseover").click();
     cy.get("#interceptors").should("not.be.empty");
-
     // add customHook to mapping step
     advancedSettingsDialog.setCustomHook("curateTile/customUriHook");
-
     advancedSettingsDialog.saveSettings(mapStep).click();
     advancedSettingsDialog.saveSettings(mapStep).should("not.be.visible");
-
-    //verify mapping step with duplicate name cannot be created
+  });
+  it("Verify mapping step with duplicate name cannot be created", () => {
     cy.waitUntil(() => curatePage.addNewStep().click());
     createEditMappingDialog.setMappingName(mapStep);
     createEditMappingDialog.setSourceRadio("Query");
@@ -192,7 +180,8 @@ describe("Mapping", () => {
     loadPage.duplicateStepErrorMessage();
     loadPage.confirmationOptions("OK").click();
     loadPage.duplicateStepErrorMessageClosed();
-
+  });
+  it("Verify link to settings, Add mapstep to existing flow, Run the flow and explore the data", () => {
     // link to settings and back
     curatePage.openSourceToEntityMap("Order", mapStep);
     cy.waitUntil(() => sourceToEntityMap.expandCollapseEntity().should("be.visible")).click();
@@ -218,8 +207,8 @@ describe("Mapping", () => {
     //Verifying the properties added by load and mapping custom hooks respectively
     cy.contains("primaryKey");
     cy.contains("uriFromCustomHook");
-
-    //Create a map step under another entity", () => {
+  });
+  it("Create a map step under another entity", () => {
     // create mapping step
     toolbar.getCurateToolbarIcon().click();
     cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
@@ -233,187 +222,5 @@ describe("Mapping", () => {
     cy.waitUntil(() => curatePage.dataPresent().should("be.visible"));
     //close modal
     cy.get("body").type("{esc}");
-  });
-
-  it("can create a load step with a custom header, can create a mapping step with a custom header, and run both steps and verify in the detail view, ", () => {
-    const flowName = "orderCustomHeaderFlow";
-    const flowName2 = "orderE2eFlow";
-    const loadStep = "loadOrderCustomHeader";
-    const mapStep = "mapOrderCustomHeader";
-    // create load step
-    toolbar.getLoadToolbarIcon().click();
-    cy.waitUntil(() => loadPage.stepName("ingestion-step").should("be.visible"));
-    loadPage.addNewButton("card").click();
-    loadPage.stepNameInput().type(loadStep);
-    loadPage.stepDescriptionInput().type("load order with a custom header");
-    loadPage.stepSourceNameInput().type("backup-ABC123");
-    loadPage.confirmationOptions("Save").click();
-    cy.findByText(loadStep).should("be.visible");
-
-    // Open step settings and switch to Advanced tab
-    loadPage.editStepInCardView(loadStep).click({force: true});
-    loadPage.switchEditAdvanced().click();
-
-    // add custom header to load step
-    advancedSettingsDialog.setHeaderContent("loadTile/customHeader");
-    advancedSettingsDialog.saveSettings(loadStep).click();
-    advancedSettingsDialog.saveSettings(loadStep).should("not.be.exist");
-
-    // add step to a new flow
-    loadPage.addStepToNewFlow(loadStep);
-    cy.findByText("New Flow").should("be.visible");
-    runPage.setFlowName(flowName);
-    runPage.setFlowDescription(`${flowName} description`);
-    loadPage.confirmationOptions("Save").click();
-    cy.verifyStepAddedToFlow("Load", loadStep);
-
-    //Run the ingest with JSON
-    runPage.runStep(loadStep).click();
-    cy.uploadFile("input/10260.json");
-    cy.verifyStepRunResult("success", "Ingestion", loadStep);
-    tiles.closeRunMessage();
-
-    // create mapping step
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    cy.waitUntil(() => curatePage.addNewStep().click());
-
-    createEditMappingDialog.setMappingName(mapStep);
-    createEditMappingDialog.setMappingDescription("An order mapping with custom header");
-    createEditMappingDialog.setSourceRadio("Query");
-    createEditMappingDialog.setQueryInput(`cts.collectionQuery(['${loadStep}'])`);
-    createEditMappingDialog.saveButton().click({force: true});
-
-    //verify that step details automatically opens after step creation
-    curatePage.verifyStepDetailsOpen(mapStep);
-    //close modal
-    cy.get("body").type("{esc}");
-
-    curatePage.verifyStepNameIsVisible(mapStep);
-
-    // Open step settings and switch to Advanced tab
-    cy.waitUntil(() => curatePage.editStep(mapStep).click({force: true}));
-    curatePage.switchEditAdvanced().click();
-
-    // add custom header
-    advancedSettingsDialog.setHeaderContent("curateTile/customHeader");
-    cy.waitUntil(() => advancedSettingsDialog.saveSettings(mapStep).click({force: true}));
-    advancedSettingsDialog.saveSettings(mapStep).should("not.exist");
-
-    // map source to entity
-    curatePage.openSourceToEntityMap("Order", mapStep);
-    cy.waitUntil(() => sourceToEntityMap.expandCollapseEntity().should("be.visible")).click();
-    sourceToEntityMap.setXpathExpressionInput("orderId", "OrderID");
-    sourceToEntityMap.setXpathExpressionInput("address", "/");
-    sourceToEntityMap.setXpathExpressionInput("city", "ShipCity");
-    sourceToEntityMap.setXpathExpressionInput("state", "ShipAddress");
-    sourceToEntityMap.setXpathExpressionInput("orderDetails", "/");
-    sourceToEntityMap.setXpathExpressionInput("productID", "OrderDetails/ProductID");
-    sourceToEntityMap.setXpathExpressionInput("unitPrice", "head(OrderDetails/UnitPrice)");
-    sourceToEntityMap.setXpathExpressionInput("quantity", "OrderDetails/Quantity");
-    sourceToEntityMap.setXpathExpressionInput("discount", "head(OrderDetails/Discount)");
-    sourceToEntityMap.setXpathExpressionInput("shipRegion", "ShipRegion");
-    sourceToEntityMap.setXpathExpressionInput("shippedDate", "ShippedDate");
-    // close modal
-    cy.get("body").type("{esc}");
-
-    //Cancel add to new flow
-    curatePage.addToNewFlow("Order", mapStep);
-    cy.findByText("New Flow").should("be.visible");
-    loadPage.confirmationOptions("Cancel").click();
-    //should route user back to curate page
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Order").should("be.visible"));
-
-    curatePage.openExistingFlowDropdown("Order", mapStep);
-    curatePage.getExistingFlowFromDropdown(flowName).click();
-    curatePage.addStepToFlowConfirmationMessage();
-    curatePage.confirmAddStepToFlow(mapStep, flowName);
-
-    runPage.runStep(mapStep).click();
-    cy.verifyStepRunResult("success", "Mapping", mapStep);
-
-    tiles.closeRunMessage();
-    runPage.deleteStep(mapStep).click();
-    loadPage.confirmationOptions("Yes").click();
-
-    //Verify Run Map step in an existing Flow
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    curatePage.runStepInCardView(mapStep).click();
-    curatePage.runStepSelectFlowConfirmation().should("be.visible");
-    curatePage.selectFlowToRunIn(flowName);
-    //Step should automatically run
-    cy.verifyStepRunResult("success", "Mapping", mapStep);
-    tiles.closeRunMessage();
-
-    //Delete the flow
-    runPage.deleteFlow(flowName).click();
-    runPage.deleteFlowConfirmationMessage(flowName).should("be.visible");
-    loadPage.confirmationOptions("Yes").click();
-    runPage.getFlowName(flowName).should("not.be.visible");
-
-    //Verify Run Map step in a new Flow
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    curatePage.runStepInCardView(mapStep).click();
-    //Just deleted flow should not be visible on flows list
-    cy.findByText(flowName).should("not.be.visible");
-    curatePage.runInNewFlow(mapStep).click({force: true});
-    cy.findByText("New Flow").should("be.visible");
-    runPage.setFlowName(flowName);
-    runPage.setFlowDescription(`${flowName} description`);
-    loadPage.confirmationOptions("Save").click();
-    //Step should automatically run
-    cy.verifyStepRunResult("success", "Mapping", mapStep);
-    tiles.closeRunMessage();
-
-    //Verify Run Map step in flow where step exists, should run automatically
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    curatePage.runStepInCardView(mapStep).click();
-    curatePage.runStepExistsOneFlowConfirmation().should("be.visible");
-    curatePage.confirmContinueRun();
-    cy.waitForAsyncRequest();
-    cy.verifyStepAddedToFlow("Map", mapStep);
-    cy.waitUntil(() => runPage.getFlowName(flowName).should("be.visible"));
-    cy.verifyStepRunResult("success", "Mapping", mapStep);
-    tiles.closeRunMessage();
-
-    // add step to a new flow
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    curatePage.addToNewFlow("Order", mapStep);
-    cy.findByText("New Flow").should("be.visible");
-    runPage.setFlowName(flowName2);
-    runPage.setFlowDescription(`${flowName2} description`);
-    loadPage.confirmationOptions("Save").click();
-    cy.verifyStepAddedToFlow("Map", mapStep);
-
-    //Verify Run Map step where step exists in multiple flows, choose one to automatically run in
-    toolbar.getCurateToolbarIcon().click();
-    cy.waitUntil(() => curatePage.getEntityTypePanel("Customer").should("be.visible"));
-    curatePage.toggleEntityTypeId("Order");
-    curatePage.runStepInCardView(mapStep).click();
-    curatePage.runStepExistsMultFlowsConfirmation().should("be.visible");
-    curatePage.selectFlowToRunIn(flowName);
-    cy.waitForAsyncRequest();
-    cy.verifyStepAddedToFlow("Map", mapStep);
-    cy.waitUntil(() => runPage.getFlowName(flowName).should("be.visible"));
-    cy.verifyStepRunResult("success", "Mapping", mapStep);
-
-    runPage.explorerLink().click();
-    browsePage.getTableViewInstanceIcon().click();
-
-    detailPage.getDocumentSource().should("contain", "backup-ABC123");
-    detailPage.getDocumentTimestamp().should("not.exist");
-
-    detailPage.getSourceView().click();
-    cy.contains("accessLevel");
-    cy.contains("999ABC");
   });
 });
