+++ marklogic-data-hub-central/ui/e2e/cypress/integration/explore/exploreTableValidations.spec.tsx
@@ -6,43 +6,33 @@ import "cypress-wait-until";
 import {toolbar} from "../../support/components/common";
 import LoginPage from "../../support/pages/login";
 
-
-describe("table test scenarios", () => {
-
-  beforeEach(() => {
+describe("Validate table and column selector in explore", () => {
+  before(() => {
     cy.visit("/");
     cy.contains(Application.title);
     cy.loginAsDeveloper().withRequest();
     LoginPage.postLogin();
+  });
+  beforeEach(() => {
+    cy.loginAsDeveloper().withRequest();
+  });
+  afterEach(() => {
+    cy.resetTestUser();
+  });
+  it("Navigate to Explore", () => {
     cy.waitUntil(() => toolbar.getExploreToolbarIcon()).click();
     cy.waitUntil(() => browsePage.getExploreButton()).click();
     browsePage.waitForSpinnerToDisappear();
     browsePage.waitForTableToLoad();
   });
-
-  it("has table and expandable rows", () => {
+  it("Validate the table and expandable rows", () => {
     browsePage.getSelectedEntity().should("contain", "All Entities");
     browsePage.getTotalDocuments().should("be.greaterThan", 25);
     browsePage.getTableRows().should("have.length", 20);
     browsePage.getTableColumns().should("be.visible");
     browsePage.getExpandable().should("be.visible");
   });
-
-});
-
-
-describe("column selector test scenarios", () => {
-
-  beforeEach(() => {
-    cy.visit("/");
-    cy.contains(Application.title);
-    cy.loginAsDeveloper().withRequest();
-    LoginPage.postLogin();
-    cy.waitUntil(() => toolbar.getExploreToolbarIcon()).click();
-    cy.waitUntil(() => browsePage.getExploreButton()).click();
-  });
-
-  it("has columns selector popover, draggable titles and checkable titles", () => {
+  it("Validate columns selector popover, draggable titles and checkable titles", () => {
     browsePage.selectEntity("Customer");
     browsePage.getSelectedEntity().should("contain", "Customer");
     browsePage.getColumnSelectorIcon().should("be.visible");
@@ -51,6 +41,5 @@ describe("column selector test scenarios", () => {
     browsePage.getTreeItemTitle(2).should("have.class", "draggable");
     browsePage.getTreeItem(2).should("have.class", "ant-tree-treenode-checkbox-checked");
   });
-
 });
