+++ marklogic-data-hub-central/ui/e2e/cypress/integration/explore/exploreXMLValidations.spec.tsx
@@ -13,17 +13,22 @@ describe("xml scenario for snippet view on browse documents page", () => {
   let facets: string[] = ["collection", "flow"];
 
   //login with valid account and go to /browse page
-  beforeEach(() => {
+  before(() => {
     cy.visit("/");
     cy.contains(Application.title);
     cy.loginAsDeveloper().withRequest();
     LoginPage.postLogin();
+  });
+  beforeEach(() => {
+    cy.loginAsDeveloper().withRequest();
+  });
+  afterEach(() => {
+    cy.resetTestUser();
+  });
+  it("select Customer XML entity instances and verify entity, docs, hub/entity properties", () => {
     cy.waitUntil(() => toolbar.getExploreToolbarIcon()).click();
     cy.waitUntil(() => browsePage.getExploreButton()).click();
     browsePage.clickFacetView();
-  });
-
-  it("select Customer XML entity instances and verify entity, docs, hub/entity properties", () => {
     browsePage.selectEntity("Customer");
     browsePage.getSelectedEntity().should("contain", "Customer");
     browsePage.getHubPropertiesExpanded();
@@ -40,13 +45,12 @@ describe("xml scenario for snippet view on browse documents page", () => {
       browsePage.getDocumentSources(i).should("exist");
       browsePage.getDocumentRecordType(i).should("exist");
     });
-
     facets.forEach(function (item) {
       browsePage.getFacet(item).should("exist");
       browsePage.getFacetItems(item).should("exist");
     });
+    browsePage.clickClearFacetSearchSelection("mapCustomersXML");
   });
-
   it("apply facet search and verify docs, hub/entity properties", () => {
     browsePage.selectEntity("All Entities");
     browsePage.getSelectedEntity().should("contain", "All Entities");
@@ -63,22 +67,22 @@ describe("xml scenario for snippet view on browse documents page", () => {
     browsePage.getFacetSearchSelectionCount("collection").should("contain", "1");
     browsePage.clickClearFacetSearchSelection("mapCustomersXML");
   });
-
   it("apply facet search and clear individual grey facet", () => {
     browsePage.selectEntity("All Entities");
     browsePage.getSelectedEntity().should("contain", "All Entities");
     browsePage.getShowMoreLink("collection").click();
     browsePage.getTotalDocuments().should("be.greaterThan", 25);
+    browsePage.showMoreCollection();
     browsePage.getFacetItemCheckbox("collection", "mapCustomersXML").click();
     browsePage.getGreySelectedFacets("mapCustomersXML").click();
     browsePage.getTotalDocuments().should("be.greaterThan", 25);
   });
-
   it("apply facet search and clear all grey facets", () => {
     browsePage.selectEntity("All Entities");
     browsePage.getSelectedEntity().should("contain", "All Entities");
     browsePage.getShowMoreLink("collection").click();
     browsePage.getTotalDocuments().should("be.greaterThan", 25);
+    browsePage.showMoreCollection();
     browsePage.getFacetItemCheckbox("collection", "Customer").click();
     browsePage.getFacetItemCheckbox("collection", "mapCustomersXML").click();
     browsePage.getGreySelectedFacets("Customer").should("exist");
@@ -86,7 +90,6 @@ describe("xml scenario for snippet view on browse documents page", () => {
     browsePage.getClearGreyFacets().click();
     browsePage.getTotalDocuments().should("be.greaterThan", 25);
   });
-
   it("search for a simple text/query and verify content", () => {
     browsePage.search("Randolph");
     browsePage.getTotalDocuments().should("be.equal", 1);
@@ -99,8 +102,9 @@ describe("xml scenario for snippet view on browse documents page", () => {
     browsePage.getDocumentRecordType(0).should("exist");
     browsePage.getDocumentRecordType(0).should("be.equal", "xml");
   });
-
   it("verify instance view of the document", () => {
+    browsePage.getSearchText().clear();
+    browsePage.waitForSpinnerToDisappear();
     browsePage.search("Randolph");
     browsePage.getTotalDocuments().should("be.equal", 1);
     browsePage.getInstanceViewIcon().click();
@@ -111,33 +115,20 @@ describe("xml scenario for snippet view on browse documents page", () => {
     detailPage.getDocumentSource().should("contain", "CustomerSourceName");
     detailPage.getDocumentRecordType().should("contain", "xml");
     detailPage.getDocumentTable().should("exist");
+    browsePage.backToResults();
+    cy.waitUntil(() => browsePage.getSearchText());
   });
-
   it("verify source view of the document", () => {
+    browsePage.getSearchText().clear();
+    browsePage.waitForSpinnerToDisappear();
     browsePage.search("Randolph");
     browsePage.getTotalDocuments().should("be.equal", 1);
     browsePage.getSourceViewIcon().click();
     detailPage.getSourceView().click();
     detailPage.getDocumentXML().should("exist");
+    browsePage.backToResults();
+    cy.waitUntil(() => browsePage.getSearchText());
   });
-
-});
-
-
-describe("xml scenario for table on browse documents page", () => {
-
-  //login with valid account and go to /browse page
-  beforeEach(() => {
-    cy.visit("/");
-    cy.contains(Application.title);
-    cy.loginAsDeveloper().withRequest();
-    LoginPage.postLogin();
-    cy.waitUntil(() => toolbar.getExploreToolbarIcon()).click();
-    cy.waitUntil(() => browsePage.getExploreButton()).click();
-    browsePage.clickTableView();
-    browsePage.selectEntity("All Entities");
-  });
-
   it("select Customer xml entity instances and verify table", () => {
     browsePage.selectEntity("Customer");
     browsePage.getSelectedEntity().should("contain", "Customer");
@@ -145,13 +136,14 @@ describe("xml scenario for table on browse documents page", () => {
     cy.waitUntil(() => browsePage.getFacetItemCheckbox("collection", "mapCustomersXML")).click();
     browsePage.getGreySelectedFacets("mapCustomersXML").should("exist");
     browsePage.getFacetApplyButton().click();
+    browsePage.clickTableView();
     browsePage.getTotalDocuments().should("be.gte", 5);
     //check table rows
     browsePage.getTableRows().should("have.length", 5);
     //check table columns
     browsePage.getTableColumns().should("have.length", 6);
+    browsePage.clickClearFacetSearchSelection("mapCustomersXML");
   });
-
   it("verify instance view of the document", () => {
     browsePage.search("Bowman");
     browsePage.getTotalDocuments().should("be.equal", 1);
@@ -163,32 +155,23 @@ describe("xml scenario for table on browse documents page", () => {
     detailPage.getDocumentSource().should("contain", "CustomerSourceName");
     detailPage.getDocumentRecordType().should("contain", "xml");
     detailPage.getDocumentTable().should("exist");
+    browsePage.backToResults();
+    cy.waitUntil(() => browsePage.getSearchText());
   });
-
   it("verify source view of the document", () => {
+    browsePage.getSearchText().clear();
+    browsePage.waitForSpinnerToDisappear();
     browsePage.search("Bowman");
     browsePage.getTotalDocuments().should("be.equal", 1);
     browsePage.getTableViewSourceIcon().click();
     detailPage.getSourceView().click();
     detailPage.getDocumentXML().should("exist");
+    browsePage.backToResults();
+    cy.waitUntil(() => browsePage.getSearchText());
+    browsePage.getSearchText().clear();
   });
-
-});
-
-describe("xml scenario for All Data option and non-entity detail page", () => {
-
-  //login with valid account and go to /browse page
-  beforeEach(() => {
-    cy.visit("/");
-    cy.contains(Application.title);
-    cy.loginAsDeveloper().withRequest();
-    LoginPage.postLogin();
-    cy.waitUntil(() => toolbar.getExploreToolbarIcon()).click();
-    cy.waitUntil(() => browsePage.getExploreButton()).click();
-    browsePage.selectEntity("All Data");
-  });
-
   it("verify record view of the XML document in non-entity detail page", () => {
+    browsePage.selectEntity("All Data");
     cy.waitUntil(() => browsePage.getNavigationIconForDocument("/dictionary/first-names.xml")).click();
     browsePage.waitForSpinnerToDisappear();
     detailPageNonEntity.getRecordView().should("exist");
