+++ marklogic-data-hub-central/ui/e2e/cypress/support/components/matching/ruleset-single-modal.tsx
@@ -9,7 +9,7 @@ class RulesetSingleModal {
   }
 
   selectMatchTypeDropdown(matchType: string) {
-    cy.findByLabelText("match-type-dropdown").should("be.visible").click();
+    cy.findByLabelText("match-type-dropdown").should("be.visible").click({force: true});
     cy.waitUntil(() => cy.findByLabelText(`${matchType}-option`).should("be.visible")).click({force: true});
   }
